+++ marshmallow/__init__.py
@@ -13,7 +13,7 @@ from marshmallow.decorators import (
 from marshmallow.utils import pprint, missing
 from marshmallow.exceptions import MarshallingError, UnmarshallingError, ValidationError
 
-__version__ = '2.0.0b5-dev'
+__version__ = '2.0.0b5'
 __author__ = 'Steven Loria'
 __license__ = 'MIT'