+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-2.0.0b5 (unreleased)
+2.0.0b5 (2015-08-23)
 ++++++++++++++++++++
 
 Features:
@@ -14,7 +14,7 @@ Features:
 Bug fixes:
 
 - `make_object` is only called after all validators and postprocessors have finished (:issue:`253`). Thanks :user:`sunsongxp` for reporting.
-- If an invalid type is passed to ``Schema`` and ``strict=False``, store a ``_schema`` error in the errors dict, rather than raising an exception (:issue:`261`). Thanks :user:`density` for reporting.
+- If an invalid type is passed to ``Schema`` and ``strict=False``, store a ``_schema`` error in the errors dict rather than raise an exception (:issue:`261`). Thanks :user:`density` for reporting.
 
 Other changes: